refactor(state representation): Tidy-up and move conversion functions (#152)
This will move the functions introduced in the previous commit to the
ParameterType.hpp header instead of the Parameter.hpp/.cpp. This makes
for a cleaner and context-specific organization.
